replaced all douubles with floats

+++ src/objectives.c
@@ -15,8 +15,8 @@
  * Sum of squares function
  * optimal solution is 0s everywhere
  */
-double sum_of_squares(const double *const args, size_t dim) {
-  double sum = 0.0;
+float sum_of_squares(const float *const args, size_t dim) {
+  float sum = 0.0;
   for (size_t idx = 0; idx < dim; idx++) {
     sum += args[idx] * args[idx];
   }
@@ -27,15 +27,15 @@ double sum_of_squares(const double *const args, size_t dim) {
  * Simple summation function
  * optimal solution is 0s everywhere
  */
-double sum(const double *const args, size_t dim) {
-  double sum = 0.0;
+float sum(const float *const args, size_t dim) {
+  float sum = 0.0;
   for (size_t idx = 0; idx < dim; idx++) {
     sum += args[idx];
   }
   return sum;
 }
 
-double sum_negative(const double *const args, size_t dim) {
+float sum_negative(const float *const args, size_t dim) {
   return -1.0 * sum(args, dim);
 }
 
@@ -43,8 +43,8 @@ double sum_negative(const double *const args, size_t dim) {
  * Rastigrin function
  * optimal solution is 0s everywhere
  */
-double rastigrin(const double *const args, size_t dim) {
-  double rstgnND = 10 * dim;
+float rastigrin(const float *const args, size_t dim) {
+  float rstgnND = 10 * dim;
   for (size_t idx = 0; idx < dim; idx++) {
     rstgnND += pow(args[dim], 2.0) - 10 * cos(2.0 * M_PI * args[dim]);
   }
@@ -55,14 +55,14 @@ double rastigrin(const double *const args, size_t dim) {
  * Multidimensional Rosenbrock Function
  * global minima f(x1,.....,xN) = 0 at (x1,.....,xN) = (1,.....,1)
  */
-double rosenbrock(const double *const args, size_t dim) {
-  double rNd = 0.;
+float rosenbrock(const float *const args, size_t dim) {
+  float rNd = 0.;
   for (size_t idx = 0; idx < dim - 1; idx++)
     rNd += (100 * (pow(args[idx + 1] - pow(args[idx], 2), 2)) + pow(1 - args[idx], 2));
   return rNd;
 }
 
-double rosenbrock_neg(const double *const args, size_t dim) {
+float rosenbrock_neg(const float *const args, size_t dim) {
   return -1.0 * rosenbrock(args, dim);
 }
 
@@ -70,8 +70,8 @@ double rosenbrock_neg(const double *const args, size_t dim) {
  * Multi Dimensional Sphere Function
  * global minima at f(x1,.....,xN) = 0 at (x1,......,xN) = (0,......,0)
  */
-double sphere(const double *const args, size_t dim) {
-  double sph = 0;
+float sphere(const float *const args, size_t dim) {
+  float sph = 0;
   for (size_t idx = 0; idx < dim; idx++)
     sph += pow(args[idx], 2);
   return sph;
@@ -81,10 +81,10 @@ double sphere(const double *const args, size_t dim) {
  * 2D Eggholder Function
  * global minima at f(x,y) = -959.6407 at (x,y) = (512,404.2319)
  */
-double egghol2d(const double *const args, size_t dim) {
+float egghol2d(const float *const args, size_t dim) {
   //assert(dim == 2);
   //assert( fabs(args[0])<=512 && fabs(args[1])<=512 );
-  double eggh = -1 * (args[1] + 47) * sin(sqrt(fabs(args[0] + (args[1] + 47))))
+  float eggh = -1 * (args[1] + 47) * sin(sqrt(fabs(args[0] + (args[1] + 47))))
                 - args[0] * sin(sqrt(fabs(args[0] - (args[1] + 47))));
   return eggh;
 }
@@ -92,8 +92,8 @@ double egghol2d(const double *const args, size_t dim) {
 /**
  * 2D Schaffer Function
  */
-double schaf2d(const double *const args, size_t dim) {
-  double sch = 0.5 + (pow(cos(pow(sin(fabs(args[0] * args[0] - args[1] * args[1])), 2)), 2) - 0.5) /
+float schaf2d(const float *const args, size_t dim) {
+  float sch = 0.5 + (pow(cos(pow(sin(fabs(args[0] * args[0] - args[1] * args[1])), 2)), 2) - 0.5) /
                      pow((1 + 0.001 * ((args[0] * args[0]) + (args[1] * args[1]))), 2);
   return sch;
 }
@@ -101,8 +101,8 @@ double schaf2d(const double *const args, size_t dim) {
 /**
  * Griewank function
  */
-double griewank(const double *const args, size_t dim) {
-  double sum = 0, prod = 1;
+float griewank(const float *const args, size_t dim) {
+  float sum = 0, prod = 1;
   for (size_t idx = 0; idx < dim; idx++) {
     sum += pow(args[idx], 2) / 4000;
     prod *= cos(args[idx] / sqrt(idx));
